requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15.0
yfinance>=0.2.40
pandas>=2.2.0
numpy>=2.0.0
transformers>=4.44.0
torch>=2.4.0
feedparser>=6.0.11
beautifulsoup4>=4.12.0
requests>=2.31.0
